create components and containers folders to organize components
