+++ src/containers/Header.tsx
@@ -6,7 +6,7 @@ import IconButton from '@material-ui/core/IconButton';
 import MailIcon from '@material-ui/icons/Mail';
 import Badge from '@material-ui/core/Badge';
 import NotificationsIcon from '@material-ui/icons/Notifications';
-import Avatar from './Avatar';
+import Avatar from '../components/Avatar';
 import EditIcon from '@material-ui/icons/Edit';
 import { createStyles, makeStyles, Theme } from '@material-ui/core/styles';
 import './Header.css';
